Reverted 37191471f712c1f9ed05140c6f42a4a2eb200d1b
Doubles are converted to integers anyway. You can't have a for loop going only half way.

+++ common/biomesoplenty/configuration/configfile/BOPConfigurationTerrainGen.java
@@ -11,7 +11,7 @@ public class BOPConfigurationTerrainGen
 {
 	public static Configuration config;
 	
-	public static double biomeSize;
+	public static int biomeSize;
 	public static boolean addToDefault;
 	public static boolean vanillaEnhanced;
 	public static boolean netherOverride;
@@ -120,7 +120,7 @@ public class BOPConfigurationTerrainGen
 		{
 			config.load();
 			
-			biomeSize = config.get("Biomes O\' Plenty World Type Settings", "Biome Size", 4.0, "Default World Type has 4. Large Biomes World Type has 6.").getDouble(4.0);
+			biomeSize = config.get("Biomes O\' Plenty World Type Settings", "Biome Size", 4, "Default World Type has 4. Large Biomes World Type has 6.").getInt(4);
 			addToDefault = config.get("Biome Settings", "Add Biomes To Default World", false).getBoolean(true);
 			vanillaEnhanced = config.get("Biome Settings", "Enhanced Vanilla Biomes", true).getBoolean(false);
 			netherOverride = config.get("Dimension Settings", "Enable Nether Override", true).getBoolean(true);
